Enhetstester og kontrollere (ok på nesten alt)

+++ WebAppOppg2/Controllers/PassengerTypeController.cs
@@ -12,7 +12,7 @@ namespace WebAppOppg2.PassengerTypeController
 {
     [ApiController]
 
-    [Route ("api/[controller]")]
+    [Route("api/[controller]")]
 
     public class PassengerTypeController : ControllerBase
     {
@@ -27,46 +27,41 @@ namespace WebAppOppg2.PassengerTypeController
             _log = log;
         }
 
-        //SaveTicket
         [HttpPost]
         public async Task<ActionResult> SavePassenger(PassengerType passengerType)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Passasjertype ikke lagret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.AddPassengerType(passengerType);
                 if (!returOK)
                 {
                     _log.LogInformation("passengertype kunne ikke lagres!");
-                    return BadRequest("");
+                    return BadRequest("Passasjertype ikke lagret, feil i DB - status: innlogget");
                 }
-                return Ok("");
+                return Ok("Passasjertype lagret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest("");
+            return BadRequest("Passasjertype ikke lagret, feil ved inputvalideringen på server - status: innlogget");
         }
 
-
-        //GetOne
         [HttpGet("{id}")]
         public async Task<ActionResult> GetOnePassengerType(int id)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Passasjertype ikke hentet, feil i DB - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 PassengerType passengerType = await _db.GetOnePassengerType(id);
                 if (passengerType == null)
                 {
                     _log.LogInformation("Fant ikke passengertype");
-                    return NotFound();
+                    return NotFound("Passasjertype ikke hentet, feil i DB - status: innlogget");
                 }
                 return Ok(passengerType);
             }
@@ -74,55 +69,52 @@ namespace WebAppOppg2.PassengerTypeController
             return BadRequest();
         }
 
-        //GetALl
         [HttpGet]
         public async Task<ActionResult> GetAllPassengerTypes()
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
+                return Unauthorized("Passasjertyper ikke hentet, feil i DB - status: ikke innlogget");
             }
-            */
             List<PassengerType> passengerTypes = await _db.GetAllPassengerTypes();
             return Ok(passengerTypes);
         }
 
-        //DeleteTicket
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeletePassengerType(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
+            {
+                return Unauthorized("Passasjertype ikke slettet, feil i DB - status: ikke innlogget");
+            }
             bool returOK = await _db.DeletePassengerType(id);
             if (!returOK)
             {
                 _log.LogInformation("Sletting av passengertype ble ikke utført");
-                return NotFound();
+                return NotFound("Passasjertype ikke slettet, feil i DB - status: innlogget");
             }
-            return Ok();
-
+            return Ok("Passasjertype slettet - status: innlogget");
         }
 
-        //EditTicket
         [HttpPut]
         public async Task<ActionResult> EditPassengerType(PassengerType passengerType)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Passasjertype ikke endret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.EditPassengerType(passengerType);
                 if (!returOK)
                 {
                     _log.LogInformation("Endringen kunne ikke utføres");
-                    return NotFound();
+                    return NotFound("Passasjertype ikke endret, feil i DB - status: innlogget");
                 }
-                return Ok();
+                return Ok("Passasjertype endret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest();
+            return BadRequest("Passasjertype ikke endret, feil ved inputvalideringen på server - status: innlogget");
         }
     }
 }
